+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 中文 | [英文](https://github.com/FightingDesign/fighting-design/blob/master/CHANGELOG.en-US.md)
 
 - `f-link` 组件 `prohibit` 配置项改为 `disabled`
+- `f-switch` 组件 `open-color` 改为 `active-color`
 
 ## 0.9.0-alpha.6 (2022-11-29)